[add] #2 번역 후, Route 연결로 원본과 OCR 텍스트 비교
